+++ scripts/transformers/webpack-5-asset-module-backport.ts
